Remove juliet-related parameters from config

+++ binary_port/src/error.rs
@@ -5,7 +5,7 @@ pub enum Error {
     #[error("Invalid request tag ({0})")]
     InvalidBinaryRequestTag(u8),
     #[error("Request too large: allowed {allowed} bytes, got {got} bytes")]
-    RequestTooLarge { allowed: usize, got: usize },
+    RequestTooLarge { allowed: u32, got: u32 },
     #[error("Empty request")]
     EmptyRequest,
     #[error(transparent)]
